pyproject.toml
[tool.poetry]
name = "tvm"
version = "0.1.0"
description = ""
authors = ["Shane Anderson <shane@reactivate.cx>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.3"
boto3 = "^1.17.80"
djangorestframework = "^3.12.4"
Markdown = "^3.3.4"
django-filter = "^2.4.0"
drf-yasg = "^1.20.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^21.5b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
